Add support for backreferencing. Fix issue with execution results. #58, #41
